+++ src/enums.ts
@@ -7,3 +7,8 @@ export enum VersionExeName {
     Stable = 'devproxy',
     Beta = 'devproxy-beta',
 }
+
+export enum WingetPackageIdentifier {
+  Stable = 'DevProxy.DevProxy',
+  Beta = 'DevProxy.DevProxy.Beta',
+}